requirements.txt
django
twilio
python-dotenv
djangorestframework
psycopg2-binary
django-redis
cryptography
dj-database-url
gunicorn
drf-spectacular
mysqlclient
